restructure lib
