+++ echo-localfiles/src/library/mod.rs
@@ -1,0 +1,7 @@
+pub mod albums;
+pub mod artists;
+pub mod tracks;
+
+pub fn get_artists() -> Vec<String> {
+    todo!()
+}
